Ensure paths.include patterns are as intended ahead of semgrep fixes (#3630)
We're aligning paths.include/exclude behavior with Semgrepignore v2
which follows the Gitignore spec. The trick is that a glob pattern that
contains a slash between two segments is now left-anchored i.e. it's the
same as if it were starting with a slash as per the Gitignore spec. In
these cases, the leading slash is optional but I'm adding it to make the
intention clearer to readers. Examples of affected patterns include `a/b`
and `a/*` but not `a/**`.
I identified three kinds of rules that had such patterns in this repo:
- Wordpress: wp-content/plugins -> **/wp-content/plugins
- Nginx: sites-available/* -> **/sites-available/*
- GitHub Actions: .github/workflows -> /.github/workflows
I'm confident that the GitHub Actions folder is only at the Git project
root. I'm not sure about Wordpress or Nginx projects, so
I'm erring on the side of caution by prepending `**/`.

Co-authored-by: Kurt Boberg <98792107+kurt-r2c@users.noreply.github.com>